feat: Refactor RPC service to improve payload handling and error publishing
