+++ girolle/src/nameko_utils/mod.rs
@@ -1,9 +1,15 @@
-use crate::error::GirolleError;
+use crate::error::{GirolleError, RemoteError};
+use crate::rpc_task::RpcTask;
+use lapin::options::BasicPublishOptions;
 /// # nameko_utils
 ///
 /// This module contains functions to manipulate the headers of the AMQP messages.
 use lapin::types::{AMQPValue, FieldTable, LongString, ShortString};
+use lapin::Channel;
 use lapin::{message::Delivery, BasicProperties};
+use serde::Deserialize;
+use serde_json::{json, Value};
+use std::collections::HashMap;
 use tracing::error;
 use uuid::Uuid;
 
@@ -26,7 +32,7 @@ fn test_set_current_call_id() {
     );
 }
 
-pub fn insert_new_id_to_call_id(
+pub(crate) fn insert_new_id_to_call_id(
     mut headers: FieldTable,
     function_name: &str,
     id: &str,
@@ -45,7 +51,7 @@ pub fn insert_new_id_to_call_id(
     headers
 }
 
-pub fn get_id(opt_id: &Option<ShortString>, id_name: &str) -> String {
+pub(crate) fn get_id(opt_id: &Option<ShortString>, id_name: &str) -> String {
     match opt_id {
         Some(id) => id.to_string(),
         None => {
@@ -60,7 +66,7 @@ fn test_get_id() {
     assert_eq!(id, "id".to_string());
 }
 
-pub fn delivery_to_message_properties(
+pub(crate) fn delivery_to_message_properties(
     delivery: &Delivery,
     id: &Uuid,
     rpc_queue: &str,
@@ -85,4 +91,183 @@ pub fn delivery_to_message_properties(
         .with_headers(headers)
         .with_delivery_mode(2)
         .with_priority(0))
+}
+
+pub(crate) async fn publish(
+    rpc_channel: &Channel,
+    payload: String,
+    properties: BasicProperties,
+    reply_to_id: String,
+    rpc_exchange: &str,
+) -> lapin::Result<()> {
+    // Need to clone the rpc_channel to be able to use it in the tokio::spawn
+    let rpc_channel_clone = rpc_channel.clone();
+    let rpc_exchange_clone = rpc_exchange.to_string();
+    tokio::spawn(async move {
+        rpc_channel_clone
+            .basic_publish(
+                &rpc_exchange_clone,
+                &format!("{}", &reply_to_id),
+                BasicPublishOptions::default(),
+                payload.as_bytes(),
+                properties,
+            )
+            .await
+            .unwrap()
+            .await
+            .unwrap();
+    });
+    // The message was correctly published
+    Ok(())
+}
+
+fn push_values_to_result(
+    service_args: &[&str],
+    kwargs: &HashMap<String, Value>,
+    start: usize,
+    end: usize,
+) -> Result<Vec<Value>, GirolleError> {
+    let mut result: Vec<Value> = Vec::new();
+    let error_message = "Key is missing in kwargs".to_string();
+    for i in start..end {
+        result.push(
+            kwargs
+                .get(service_args[i])
+                .ok_or_else(|| GirolleError::IncorrectSignature(error_message.clone()))?
+                .clone(),
+        );
+    }
+    Ok(result)
+}
+
+fn build_inputs_fn_service(
+    service_args: &Vec<&str>,
+    data_delivery: DeliveryData,
+) -> Result<Vec<Value>, GirolleError> {
+    let args_size: usize = data_delivery.args.len();
+    let kwargs_size: usize = data_delivery.kwargs.len();
+    let service_args_size: usize = service_args.len();
+
+    match (
+        data_delivery.kwargs.is_empty(),
+        data_delivery.args.is_empty(),
+        service_args_size == args_size + kwargs_size,
+    ) {
+        (true, _, _) if service_args_size == args_size => Ok(data_delivery.args),
+        (_, true, _) if service_args_size == kwargs_size => {
+            push_values_to_result(service_args, &data_delivery.kwargs, 0, kwargs_size)
+        }
+        (_, _, true) => {
+            let mut result = data_delivery.args;
+            result.extend(push_values_to_result(
+                service_args,
+                &data_delivery.kwargs,
+                args_size,
+                args_size + kwargs_size,
+            )?);
+            Ok(result)
+        }
+        _ => Err(GirolleError::IncorrectSignature(format!(
+            "takes {} positional arguments but {} were given",
+            service_args_size,
+            args_size + kwargs_size
+        ))),
+    }
+}
+
+#[test]
+fn test_build_inputs_fn_service() {
+    let service_args = vec!["a", "b", "c"];
+    let data_delivery = DeliveryData {
+        args: vec![
+            Value::String("1".to_string()),
+            Value::String("2".to_string()),
+        ],
+        kwargs: HashMap::from([("c".to_string(), Value::String("3".to_string()))]),
+    };
+    let result = build_inputs_fn_service(&service_args, data_delivery);
+    assert_eq!(result.is_ok(), true);
+    let result = result.unwrap();
+    assert_eq!(result.len(), 3);
+    assert_eq!(result[0], Value::String("1".to_string()));
+    assert_eq!(result[1], Value::String("2".to_string()));
+    assert_eq!(result[2], Value::String("3".to_string()));
+}
+
+fn get_result_paylaod(result: Value) -> String {
+    json!(
+        {
+            "result": result,
+            "error": null,
+        }
+    )
+    .to_string()
+}
+
+pub(crate) fn get_error_payload(error: RemoteError) -> String {
+    json!(
+        {
+            "result": null,
+            "error": error,
+        }
+    )
+    .to_string()
+}
+#[derive(Debug, Deserialize)]
+pub struct DeliveryData {
+    args: Vec<Value>,
+    kwargs: HashMap<String, Value>,
+}
+/// Execute the delivery
+pub(crate) async fn compute_deliver(
+    incomming_data: DeliveryData,
+    properties: BasicProperties,
+    rpc_task_struct: &RpcTask,
+    rpc_channel: &Channel,
+    rpc_exchange: &str,
+    reply_to_id: String,
+) {
+    // Publish the response
+    let fn_service = rpc_task_struct.inner_function;
+    let buildted_args = match build_inputs_fn_service(&rpc_task_struct.args, incomming_data) {
+        Ok(result) => result,
+        Err(error) => {
+            publish(
+                &rpc_channel,
+                get_error_payload(error.convert()),
+                properties,
+                reply_to_id,
+                rpc_exchange,
+            )
+            .await
+            .expect("Error publishing");
+            return;
+        }
+    };
+    match fn_service(&buildted_args) {
+        Ok(result) => {
+            publish(
+                &rpc_channel,
+                get_result_paylaod(result),
+                properties,
+                reply_to_id,
+                rpc_exchange,
+            )
+            .await
+            .expect("Error publishing");
+            return;
+        }
+        Err(error) => {
+            publish(
+                &rpc_channel,
+                get_error_payload(error.convert()),
+                properties,
+                reply_to_id,
+                rpc_exchange,
+            )
+            .await
+            .expect("Error publishing");
+            return;
+        }
+    };
 }
