+++ girolle/src/rpc_service.rs
@@ -1,13 +1,14 @@
 use crate::{
     config::Config,
-    error::{GirolleError, RemoteError},
-    nameko_utils::{delivery_to_message_properties, get_id},
+    error::GirolleError,
+    nameko_utils::{
+        compute_deliver, delivery_to_message_properties, get_error_payload, get_id, publish,
+        DeliveryData,
+    },
     queue::{create_service_channel, get_connection},
     rpc_task::RpcTask,
 };
-use lapin::{message::DeliveryResult, options::*, types::FieldTable, BasicProperties, Channel};
-use serde::Deserialize;
-use serde_json::{json, Value};
+use lapin::{message::DeliveryResult, options::*, types::FieldTable, Channel};
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::Semaphore;
 use tracing::{debug, error, info, warn, Level};
@@ -221,185 +222,6 @@ impl RpcService {
         self.conf = config;
         Ok(())
     }
-}
-
-async fn publish(
-    rpc_channel: &Channel,
-    payload: String,
-    properties: BasicProperties,
-    reply_to_id: String,
-    rpc_exchange: &str,
-) -> lapin::Result<()> {
-    // Need to clone the rpc_channel to be able to use it in the tokio::spawn
-    let rpc_channel_clone = rpc_channel.clone();
-    let rpc_exchange_clone = rpc_exchange.to_string();
-    tokio::spawn(async move {
-        rpc_channel_clone
-            .basic_publish(
-                &rpc_exchange_clone,
-                &format!("{}", &reply_to_id),
-                BasicPublishOptions::default(),
-                payload.as_bytes(),
-                properties,
-            )
-            .await
-            .unwrap()
-            .await
-            .unwrap();
-    });
-    // The message was correctly published
-    Ok(())
-}
-
-fn push_values_to_result(
-    service_args: &[&str],
-    kwargs: &HashMap<String, Value>,
-    start: usize,
-    end: usize,
-) -> Result<Vec<Value>, GirolleError> {
-    let mut result: Vec<Value> = Vec::new();
-    let error_message = "Key is missing in kwargs".to_string();
-    for i in start..end {
-        result.push(
-            kwargs
-                .get(service_args[i])
-                .ok_or_else(|| GirolleError::IncorrectSignature(error_message.clone()))?
-                .clone(),
-        );
-    }
-    Ok(result)
-}
-
-fn build_inputs_fn_service(
-    service_args: &Vec<&str>,
-    data_delivery: DeliveryData,
-) -> Result<Vec<Value>, GirolleError> {
-    let args_size: usize = data_delivery.args.len();
-    let kwargs_size: usize = data_delivery.kwargs.len();
-    let service_args_size: usize = service_args.len();
-
-    match (
-        data_delivery.kwargs.is_empty(),
-        data_delivery.args.is_empty(),
-        service_args_size == args_size + kwargs_size,
-    ) {
-        (true, _, _) if service_args_size == args_size => Ok(data_delivery.args),
-        (_, true, _) if service_args_size == kwargs_size => {
-            push_values_to_result(service_args, &data_delivery.kwargs, 0, kwargs_size)
-        }
-        (_, _, true) => {
-            let mut result = data_delivery.args;
-            result.extend(push_values_to_result(
-                service_args,
-                &data_delivery.kwargs,
-                args_size,
-                args_size + kwargs_size,
-            )?);
-            Ok(result)
-        }
-        _ => Err(GirolleError::IncorrectSignature(format!(
-            "takes {} positional arguments but {} were given",
-            service_args_size,
-            args_size + kwargs_size
-        ))),
-    }
-}
-
-#[test]
-fn test_build_inputs_fn_service() {
-    let service_args = vec!["a", "b", "c"];
-    let data_delivery = DeliveryData {
-        args: vec![
-            Value::String("1".to_string()),
-            Value::String("2".to_string()),
-        ],
-        kwargs: HashMap::from([("c".to_string(), Value::String("3".to_string()))]),
-    };
-    let result = build_inputs_fn_service(&service_args, data_delivery);
-    assert_eq!(result.is_ok(), true);
-    let result = result.unwrap();
-    assert_eq!(result.len(), 3);
-    assert_eq!(result[0], Value::String("1".to_string()));
-    assert_eq!(result[1], Value::String("2".to_string()));
-    assert_eq!(result[2], Value::String("3".to_string()));
-}
-
-fn get_result_paylaod(result: Value) -> String {
-    json!(
-        {
-            "result": result,
-            "error": null,
-        }
-    )
-    .to_string()
-}
-
-fn get_error_payload(error: RemoteError) -> String {
-    json!(
-        {
-            "result": null,
-            "error": error,
-        }
-    )
-    .to_string()
-}
-#[derive(Debug, Deserialize)]
-struct DeliveryData {
-    args: Vec<Value>,
-    kwargs: HashMap<String, Value>,
-}
-/// Execute the delivery
-async fn compute_deliver(
-    incomming_data: DeliveryData,
-    properties: BasicProperties,
-    rpc_task_struct: &RpcTask,
-    rpc_channel: &Channel,
-    rpc_exchange: &str,
-    reply_to_id: String,
-) {
-    // Publish the response
-    let fn_service = rpc_task_struct.inner_function;
-    let buildted_args = match build_inputs_fn_service(&rpc_task_struct.args, incomming_data) {
-        Ok(result) => result,
-        Err(error) => {
-            publish(
-                &rpc_channel,
-                get_error_payload(error.convert()),
-                properties,
-                reply_to_id,
-                rpc_exchange,
-            )
-            .await
-            .expect("Error publishing");
-            return;
-        }
-    };
-    match fn_service(&buildted_args) {
-        Ok(result) => {
-            publish(
-                &rpc_channel,
-                get_result_paylaod(result),
-                properties,
-                reply_to_id,
-                rpc_exchange,
-            )
-            .await
-            .expect("Error publishing");
-            return;
-        }
-        Err(error) => {
-            publish(
-                &rpc_channel,
-                get_error_payload(error.convert()),
-                properties,
-                reply_to_id,
-                rpc_exchange,
-            )
-            .await
-            .expect("Error publishing");
-            return;
-        }
-    };
 }
 
 struct SharedData {
